updated the dynamic rates projects

+++ Dynamic Branch/IMS_PowerDept/AppCode/NewProperties.cs
@@ -7,8 +7,8 @@ namespace IMS_PowerDept.AppCode
 {
     public class NewProperties
     {
-        protected int ChallanNO;
-        public int challanNO
+        protected double ChallanNO;
+        public double  challanNO
         {
             get { return ChallanNO; }
             set { ChallanNO = value; }
